provide a separate standalone version for prettier plugin
