+++ packages/prettier-plugin/src/mock-cjs.cjs
@@ -1,0 +1,3 @@
+// this plugin doesn't work with Prettier 2 (which we have to use for VS Code),
+// but Prettier still tries to load it, so we provide a mock version.
+exports.languages = [];
